UCH: the Urine Citrate-Hippurate metabolomic dataset

This dataset comes from a 1H NMR analysis of urine of female rats with hippuric and citric acid were added to the samples in different known concentrations.

Usage

data("UCH")

Format

A list of length 3: the experimental design (`design`), the outcomes for every observations (`outcomes`) and the formula considered to analyze the data (`formula`).

design

A data.frame with the experimental design of 34 observations and 5 explanatory variables: Hippurate: concentration of hippuric acid; Citrate: concentration of citric acid; Dilution: dilution, here all the samples are diluted with a dilution rate of 50 %; Day: for each medium, the preparation of the mixtures were performed in two series; Time: each mixture or experimental condition was repeated twice.

outcomes

A numerical matrix with 34 observations and 600 response variables

formula

A character string with the suggested formula to analyze the data

Details

The UCH vignette can be accessed with: (print(vignette(topic = "UCH", package = "limpca"))).

The database has been experimentally created in order to control the spectral locations of the biomarkers to find (i.e. Hippurate and Citrate). This property allows us to evaluate the performances of the data analysis of various statistical methods. This urine experimental database is also designed in order to explore the influence on spectra of intra-sample 1H NMR replications (Time), and inter-day 1H NMR measurements (Day).

The model formula is:

outcomes = Hippurate + Citrate + Time + Hippurate:Citrate + Time:Hippurate + Time:Citrate + Hippurate:Citrate:Time

Source

Martin, M. (2020). Uncovering informative content in metabolomics data: from pre-processing of 1H NMR spectra to biomarkers discovery in multifactorial designs. Prom.: Govaerts, B. PhD thesis. Institut de statistique, biostatistique et sciences actuarielles, UCLouvain, Belgium. http://hdl.handle.net/2078.1/227671

Examples

data("UCH")
str(UCH)
#> List of 3
#>  $ design  :'data.frame':	34 obs. of  5 variables:
#>   ..$ Hippurate: Factor w/ 3 levels "0","1","2": 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ Citrate  : Factor w/ 3 levels "0","2","4": 1 1 2 2 3 3 1 1 2 2 ...
#>   ..$ Dilution : Factor w/ 1 level "diluted": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Day      : Factor w/ 2 levels "2","3": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ Time     : Factor w/ 2 levels "1","2": 1 2 1 2 1 2 1 2 1 2 ...
#>  $ outcomes: num [1:34, 1:600] 0.0312 0.0581 0.027 0.0341 0.0406 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$   : chr [1:34] "M2C00D2R1" "M2C00D2R2" "M2C02D2R1" "M2C02D2R2" ...
#>   .. ..$ X1: chr [1:600] "9.9917004" "9.9753204" "9.9590624" "9.9427436" ...
#>  $ formula : chr "outcomes ~ Hippurate + Citrate + Time + Hippurate:Citrate + Time:Hippurate + Time:Citrate + Hippurate:Citrate:Time"