Converts data to a lmpDataList list.

Creates the lmpDataList list from a SummarizedExperiment or by manually defining the design, the outcomes and the model formula. lmpDataList serves as an input for the lmpModelMatrix function to start the limpca modeling.

Usage

data2lmpDataList(
  se = NULL,
  assay_name = NULL,
  outcomes = NULL,
  design = NULL,
  formula = NULL,
  verbose = TRUE
)

Arguments

se

A SummarizedExperiment object.

assay_name

If not NULL (default), a character string naming the assay from the SummarizedExperiment object se. If NULL, the first assay is selected.

outcomes

If not NULL (default), a numerical matrix with n observations and m response variables. The rownames needs to be non-NULL and match those of the design matrix.

design

If not NULL (default), a data.frame with the experimental design of n observations and q explanatory variables. The rownames of design has to match the rownames of outcomes.

formula

If not NULL (default), a character string with the formula that will be used to analyze the data. Only the right part of the formula is necessary, eg: "~ A + B", The names of the formula should marcht the column names of the design

verbose

If TRUE, prints useful information about the outputted list.

Value

A list with the 3 following named elements:

outcomes

A nxm matrix with the m response variables.

design

A nxq data.frame with the experimental design.

formula

A character string with the model formula.

Details

Data can be included as a SummarizedExperiment (SE) object or by manually defining one or multiple elements of outcomes, design and formula. If a SE is provided, the outcomes corresponds to a transposed assay of the SE (by default the first one), the design corresponds to the colData of the SE and the formula can be provided as a formula element in the S4Vectors::metadata of SE (metadata(se)$formula).

In the outputted list, the outcomes are structured in a standard statistical fashion, i.e. with observations in rows and the variables (features) in column. If the outcomes argument is not NULL, it has to be formatted that way (see Arguments).

Note that there is a priority to the outcomes, design and formula arguments if they are not NULL (e.g. if both se and outcomes arguments are provided, the resulting outcomes matrix will be from the outcomes argument).

Multiple checks are performed to ensure that the data are correctly formatted:

the rownames of design and outcomes should match

the names of the model terms in the formula should match column names from the design

See also

Examples

data(UCH)

### create manually the dataset

res <- data2lmpDataList(
  outcomes = UCH$outcomes,
  design = UCH$design[, 1, drop = FALSE], formula = "~ Hippurate"
)
#> | dim outcomes: 34x600
#> | formula: ~ Hippurate
#> | design variables (1): Hippurate (factor)

### create the dataset from a SummarizedExperiment

se_UCH <- SummarizedExperiment(
  assays = list(
    counts = t(UCH$outcomes),
    counts2 = t(UCH$outcomes * 2)
  ), colData = UCH$design,
  metadata = list(formula = "~ Hippurate + Citrate")
)
#> Error in SummarizedExperiment(assays = list(counts = t(UCH$outcomes),     counts2 = t(UCH$outcomes * 2)), colData = UCH$design, metadata = list(formula = "~ Hippurate + Citrate")): could not find function "SummarizedExperiment"

res <- data2lmpDataList(se_UCH, assay_name = "counts2")
#> Error in eval(expr, envir, enclos): object 'se_UCH' not found

# changing the formula:
res <- data2lmpDataList(se_UCH,
  assay_name = "counts2",
  formula = "~ Hippurate + Citrate + Time"
)
#> Error in eval(expr, envir, enclos): object 'se_UCH' not found